More renderTexture testing and fixes plus a couple of other things.
Added __debugbreak() to lmAssert for easier debugging.
Changed the gfx texture id to include an incrementing version / id so we can detect when textures get reused unintentionally due to a new texture getting the id of a disposed one. Now every new texture has a unique id, but the lower 12 bits still act as an index.
Implemented RenderTexture.clear